Introduce domain ports, slice GameContext, and split boss_system
Phase 1 ports (src/ports/): AudioPort, JuicePort, HudPort, InventoryPort,
CollisionPort, and PlayerMotionPort. Domain modules now depend on narrow
interfaces instead of singletons:

- collectibles: AudioPort injected into OrbManager
- resource_harvester: InventoryPort instead of SaveManager
- sling_combo: JuicePort + HudPort + AudioPort in options
- obstacle_system: CollisionPort types for WASM handles
- hud_manager: audio injected via constructor (no getAudioSystem)

Phase 2: GameContext field groups (CoreRuntime, FrameCounters, RunState,
GameContextExtensions) documented in game_runtime.ts and GAME_CONTEXT.md.

Phase 3: boss_system.ts split into boss_system/ folder (pitcher minion,
star eater boss, boss manager, types).

All gates green: npm run check, build, smoke.
