+++ src/boss_system/index.ts
@@ -1,0 +1,5 @@
+export type { BossPhase, BossConfig, BossHitboxEntry } from './types';
+export { BOSS_DISPLAY_NAME, PHASE_NAMES } from './types';
+export { PitcherMinion } from './pitcher_minion';
+export { StarEaterBoss } from './star_eater_boss';
+export { BossManager } from './boss_manager';
