+++ src/ports/hud_port.ts
@@ -1,0 +1,6 @@
+/** HUD surfaces used by combat feedback and sling combo. */
+export interface HudPort {
+    showSlingCombo(combo: number, isSurge?: boolean): void;
+    hideSlingCombo(): void;
+    addScore?(points: number): void;
+}
